[permissions] Divide and conquer
