+++ shared/src/commonMain/kotlin/ch/dreipol/multiplatform/reduxsample/shared/redux/reducer/SettingsReducer.kt
@@ -1,0 +1,27 @@
+package ch.dreipol.multiplatform.reduxsample.shared.redux.reducer
+
+import ch.dreipol.multiplatform.reduxsample.shared.redux.NullableState
+import ch.dreipol.multiplatform.reduxsample.shared.redux.SettingsState
+import ch.dreipol.multiplatform.reduxsample.shared.redux.actions.NextRemindersCalculated
+import ch.dreipol.multiplatform.reduxsample.shared.redux.actions.SettingsInitializedAction
+import ch.dreipol.multiplatform.reduxsample.shared.redux.actions.SettingsLoadedAction
+import org.reduxkotlin.Reducer
+
+val settingsReducer: Reducer<NullableState<SettingsState>> = { state, action ->
+    val settingsState = state.state
+    when (action) {
+        is SettingsInitializedAction -> NullableState(
+            SettingsState(
+                action.settings,
+                action.notificationPermission,
+                action.notificationSettings,
+                action.nextReminders
+            )
+        )
+        is SettingsLoadedAction -> NullableState(
+            settingsState!!.copy(settings = action.settings, notificationSettings = action.notificationSettings)
+        )
+        is NextRemindersCalculated -> NullableState(settingsState!!.copy(nextReminders = action.nextReminders))
+        else -> state
+    }
+}
